вынесены enum в отдельный файл в users/enums.ts из create.users.dto.ts и в users.entity.ts
