+++ src/users/enums.ts
@@ -1,0 +1,9 @@
+export enum Gender {
+    MALE = 'М',
+    FEMALE = 'Ж',
+}
+
+export enum UserRole {
+    USER = 'user',
+    ADMIN = 'admin',
+}
